requirements-cu128.txt
--index-url https://download.pytorch.org/whl/cu128
--extra-index-url https://pypi.org/simple

torch==2.11.0
torchvision==0.26.0
torchaudio==2.11.0
timm==1.0.26
tqdm==4.67.1
numpy==2.2.6
Pillow==11.3.0
